app/plug-in/plug-in-params.c libgimp/gimp.[ch]
2006-06-25  Manish Singh  <yosh@gimp.org>

        * app/plug-in/plug-in-params.c
        * libgimp/gimp.[ch]
        * libgimpbase/gimpprotocol.[ch]
        * plug-ins/common/colormap-remap.c
        * plug-ins/common/curve_bend.c
        * plug-ins/common/grid.c
        * plug-ins/pygimp/pygimp-pdb.c
        * plug-ins/script-fu/siod-wrapper.c
        * tools/pdbgen/pdb.pl: turn d_int8 and d_int8array into guint8.
        Fixes bug #344431.

+++ plug-ins/common/grid.c
@@ -189,7 +189,7 @@ run (const gchar      *name,
           grid_cfg.hcolor  = param[6].data.d_color;
 
           gimp_rgb_set_alpha (&(grid_cfg.hcolor),
-                              ((double) (guint) param[7].data.d_int8) / 255.0);
+                              ((double) param[7].data.d_int8) / 255.0);
 
 
           grid_cfg.vwidth  = MAX (0, param[8].data.d_int32);
@@ -198,7 +198,7 @@ run (const gchar      *name,
           grid_cfg.vcolor  = param[11].data.d_color;
 
           gimp_rgb_set_alpha (&(grid_cfg.vcolor),
-                              ((double) (guint)  param[12].data.d_int8) / 255.0);
+                              ((double) param[12].data.d_int8) / 255.0);
